Move helper object to different package
